requirements.txt
Pillow==8.1.2
pyzbar==0.1.8
